% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cua_base_discounted}
\alias{cua_base_discounted}
\title{Example cost-effectiveness dataset: cua_base_discounted}
\format{
Same structure as \code{cua_base}.
}
\source{
Synthetic example generated for internal demonstrations.
}
\usage{
cua_base_discounted
}
\description{
Discounted version of the \code{cua_base} dataset. Costs and effectiveness
values have been adjusted using a discount rate for demonstration purposes.
}
\keyword{datasets}
