% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cua_multi_discounted}
\alias{cua_multi_discounted}
\title{Multigroup cost-effectiveness dataset: cua_multi_discounted}
\format{
Same structure as \code{cua_multi}.
}
\source{
Synthetic example generated for internal demonstrations.
}
\usage{
cua_multi_discounted
}
\description{
Discounted version of the \code{cua_multi} dataset. All cost and effectiveness
values have been adjusted using a discount rate to illustrate scenario-based
or sensitivity analyses.
}
\keyword{datasets}
