% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cea_results_list.R
\name{plot.cea_results_list}
\alias{plot.cea_results_list}
\title{Plot method for cea_results_list objects}
\usage{
\method{plot}{cea_results_list}(x, type = "icers", ...)
}
\arguments{
\item{x}{An object of class \code{cea_results_list}.}

\item{type}{Character. One of "icers", "ceacs", "evpis", or "marginals".
If not specified, defaults to "icers".}

\item{...}{Additional arguments passed to the corresponding plotting method.}
}
\value{
A ggplot object.
}
\description{
Generic dispatcher for visualizations of cost-effectiveness analyses.
Redirects the call to the appropriate specific plotting function
(ICER, CEAC, EVPI, or Marginals) depending on the arguments provided.
}
