% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAgroupRui.R
\name{CEAgroupRui}
\alias{CEAgroupRui}
\title{Launch interactive CEAgroupR Shiny interface (refined classic UI)}
\usage{
CEAgroupRui()
}
\description{
Interactive Shiny application for performing and visualizing
multistrategy cost-effectiveness analyses using the CEAgroupR package.
}
