% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cua_base.R
\docType{data}
\name{cua_base}
\alias{cua_base}
\title{Example dataset for cost-utility analysis (base case)}
\format{
A tibble with 200 observations and 7 variables:
...
}
\usage{
cua_base
}
\description{
Simulated dataset representing a simple scenario for incremental cost-utility
analysis between two treatment groups (g0 and g1). It includes total cost,
a single measure of effectiveness (QALYs or utilities), and several subgroup
variables for stratified analyses.
}
\keyword{datasets}
