% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cea_base.R
\name{print.cea_base}
\alias{print.cea_base}
\title{Print method for cea_base objects}
\usage{
\method{print}{cea_base}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cea_base}.}

\item{...}{Additional arguments (not used).}
}
\description{
Provides a concise and clean console output for \code{cea_base} objects.
This print method hides the internal \code{boot} call and dataset details,
displaying only the relevant summary of the bootstrap analysis.
}
