% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_stats.R
\name{print.summary_stats}
\alias{print.summary_stats}
\title{Print method for summary statistics of CEA datasets}
\usage{
\method{print}{summary_stats}(x, digits = 3, ...)
}
\arguments{
\item{x}{A tibble produced by \code{\link{summarize_cea_data}}.}

\item{digits}{Integer. Number of decimal places to display (default = 3).}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Formats and prints descriptive summaries produced by
\code{\link{summarize_cea_data}}. Compatible with the output of
\code{\link{compute_icers}}, displaying also stored analysis settings
if available.
}
